[tests] Share common code between test runners
