+++ expression-test/test_runner_common.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <mbgl/util/feature.hpp>
+
+#include <string>
+#include <vector>
+
+using namespace mbgl;
+
+Value stripPrecision(const Value& value);
+std::vector<std::string> tokenize(std::string str);
+bool deepEqual(const Value& a, const Value& b);
+bool deepEqual(const optional<Value>& a, const optional<Value>& b);
